requirements.txt
Flask==3.0.0
tensorflow==2.18.0
keras>=3.5.0
Pillow==10.0.1
numpy==1.26.2
gunicorn==20.1.0
Werkzeug==3.0.0
